// src/app/models/review.ts
export class Review {
  id :any;
  title: string;
  message: string;
  rating: any;
  user_id: number;
  status: number;
  name: string;
  email: string;
  strain_id: number;
  rating_apperance: any;
  rating_aroma: any;
  rating_high: any;
  rating_quality: any;
  rating_flavor: any;
  rating_value: any;
}